updated the script

+++ installer/tools/installer.ps1
@@ -22,7 +22,6 @@ if (-not $msiUrl ) {
     # if the tag has "-"
     $majorVersion = $version -split "-"
     $majorVersion = $majorVersion[0]
-    echo $majorVersion
     $msiUrl = "https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/v${version}/sensing-dev-installer-${majorVersion}-win64.msi"
 }
 
@@ -32,7 +31,9 @@ Invoke-WebRequest -Uri $msiUrl -OutFile $tempMsiPath -Verbose
 
 # Install MSI to specified path
 # Note: This assumes the MSI accepts TARGETDIR as an argument for installation directory. Some MSIs might not.
-Start-Process -Wait -FilePath "msiexec.exe" -ArgumentList "/i `"$tempMsiPath`" TARGETDIR=`"$installPath`" /qn"
+
+Start-Process -Wait -FilePath "msiexec.exe" -ArgumentList "/i `"$tempMsiPath`" INSTALL_ROOT=`"$installPath`" /qn"
+
 
 # Run .ps1 file from the installed package
 $ps1ScriptPath = Join-Path -Path $installPath -ChildPath $relativeScriptPath
@@ -43,4 +44,6 @@ if (Test-Path -Path $ps1ScriptPath -PathType Leaf) {
 }
 
 # Cleanup
-Remove-Item -Path $tempMsiPath -Force
+if (Test-Path -Path $tempMsiPath ){
+    Remove-Item -Path $tempMsiPath -Force
+}
